refactor(models): share valid name generation
